Added some logging of resource init failures renamed stage0 to sizeAndStageCount removed transitionFromRenderTarget and added transition to DMADest, ShaderSrc and DMASrc Vulkan Validation showed upA lots of errors, fixed them Added support for debug marker names but apparently extension not available... Added support for required and desired vulkan instances.

+++ libs/resourcemanager/resourceman.h
@@ -44,7 +44,7 @@ using HasResourceChangedFunc = std::function<bool(ResourceBase::ConstPtr ptr_)>;
 using SaveResourceFunc = std::function<bool(ResourceBase::ConstPtr ptr_, ISaver& saver_)>;
 
 // extra memory, init, destroy
-using ResourceHandler = std::tuple<int, HandlerInit, HandlerDestroy>;
+using ResourceHandler = std::tuple<size_t, HandlerInit, HandlerDestroy>;
 
 class ResourceMan
 {
